api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.2
python-jose[cryptography]==3.4.0
passlib==1.7.4
bcrypt>=3.2.0,<4.1.0
python-multipart==0.0.18
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
redis==5.0.1
prometheus-client==0.19.0
pytest==7.4.3
httpx==0.25.2
requests==2.32.0
Pillow>=10.0.0
structlog==24.1.0
colorama==0.4.6
psutil==6.0.0